🛠 🤷‍♂️ course & quizzes: attempt at refactoring util fxs
